% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b_anomaly}
\alias{b_anomaly}
\title{Example data computed for B anomaly observables.}
\format{
A list with four entries
}
\usage{
b_anomaly
}
\description{
A list containing predictions (pred), parameters (wc), inverse covariance
matrix (covInv) and observed values (exp) for the parameters C9 and C10 and
a set of neutral B decay observables.
}
\references{
TODO
}
\keyword{datasets}
