% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bikes}
\alias{Bikes}
\title{bikes with activations}
\format{
a list  of 2 dataframes
\describe{
\item{space1}{dataframe 731 obs of 10 variables (cluster space)}
\item{space2}{dataframe 731 obs of 6 variables (space2)}
}
}
\usage{
Bikes
}
\description{
data from Ursula
}
\keyword{datasets}
