% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{chi2Bins}
\alias{chi2Bins}
\title{Bin points based on chi2}
\usage{
chi2Bins(chivals, ndf, k)
}
\arguments{
\item{chivals}{vector with chi2 values}

\item{ndf}{number of parameters (degrees of freedom of the chi2 distribution)}

\item{k}{number of bins}
}
\value{
bin assignment for each point
}
\description{
Map to values of sigma and compute equidistant binning in sigma.
}
\keyword{internal}
