% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{colourHelper}
\alias{colourHelper}
\title{function for assigning colouring, palette and labels}
\usage{
colourHelper(choice, rv)
}
\arguments{
\item{choice}{choice of colouring for an output}

\item{rv}{reactive variables}
}
\value{
list containing colour assignment, palette and labels for use in plotting
}
\description{
function for assigning colouring, palette and labels
}
\keyword{internal}
