% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{computeChi2}
\alias{computeChi2}
\title{Compute chi2 value for all points}
\usage{
computeChi2(pred, covInv, exp)
}
\arguments{
\item{pred}{matrix of predicted values for all points}

\item{covInv}{inverse covariance matrix}

\item{exp}{experimentally observed values}
}
\value{
vector with chi2 values
}
\description{
Compute chi2 value for all points
}
