% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{computeSigma}
\alias{computeSigma}
\title{Compute sigma}
\usage{
computeSigma(chivals, ndf)
}
\arguments{
\item{chivals}{vector with chi2 values}

\item{ndf}{number of parameters (degrees of freedom of the chi2 distribution)}
}
\value{
vector with sigma values
}
\description{
Map chi2 to sigma, with cutoff (overflow) at 5 sigma
}
