% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cstat_labeller}
\alias{cstat_labeller}
\title{function for labeling cluster statistics on statistics page of pandemonium GUI}
\usage{
cstat_labeller()
}
\description{
function for labeling cluster statistics on statistics page of pandemonium GUI
}
\keyword{internal}
