% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getClusterStats}
\alias{getClusterStats}
\title{Compute cluster statistics}
\usage{
getClusterStats(dist, fit, chivals, kmax = 10)
}
\arguments{
\item{dist}{distances}

\item{fit}{result from hclust}

\item{chivals}{vector with chi2 values}

\item{kmax}{maximum number of clusters considered}
}
\value{
data frame with cluster statistics
}
\description{
For number of clusters k between two and kmax, evaluate cluster
statistics collected in output tibble.
}
