% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getCoords}
\alias{getCoords}
\title{Compute coordinates for all points}
\usage{
getCoords(coord, useCov, pred, covInv, exp, user_coord = NULL)
}
\arguments{
\item{coord}{name of coordinates to be computed}

\item{useCov}{should correlation be included in the coordinates}

\item{pred}{matrix of predicted values for all points}

\item{covInv}{inverse covariance matrix}

\item{exp}{experimentally observed values}

\item{user_coord}{user coordinates returned with coord=User}
}
\value{
matrix with coordinate representation of all points
}
\description{
Compute coordinates for all points
}
