% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getDists}
\alias{getDists}
\title{Compute distances between all points}
\usage{
getDists(coord, metric, user_dist = NULL)
}
\arguments{
\item{coord}{matrix with coordinate representation of all points}

\item{metric}{name of distance metric to be used}

\item{user_dist}{user distance returned with metric=user}
}
\value{
distances between all points
}
\description{
Compute distances between all points
}
