% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{launchApp}
\alias{launchApp}
\title{Shiny app for exploring clustering solutions}
\usage{
launchApp(pred, covInv, wc, exp, user_coord = NULL, user_dist = NULL)
}
\arguments{
\item{pred}{prediction matrix}

\item{covInv}{inverse covariance matrix}

\item{wc}{matrix specifying the model parameters (on a grid)}

\item{exp}{observable reference value (e.g. experimental measurement)}

\item{user_coord}{input coordinate matrix (optional)}

\item{user_dist}{input distance matrix (optional)}
}
\description{
Opening the GUI to cluster the data points based on the predictions.
Coordinates and distances are computed on the fly, or can be entered
in the function call.
}
\examples{
\dontrun{
launchApp(b_anomaly$pred, b_anomaly$covInv, b_anomaly$wc, b_anomaly$exp)
}
}
