% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{makePlots}
\alias{makePlots}
\title{Generate a specified plot outside the GUI}
\usage{
makePlots(
  cluster,
  settings,
  cov = NULL,
  covInv = NULL,
  exp = NULL,
  linked = NULL,
  linked.cov = NULL,
  linked.covInv = NULL,
  linked.exp = NULL,
  user_dist = NULL,
  getCoordsSpace1 = normCoords,
  getCoordsSpace2 = normCoords,
  getScore = NULL,
  results = NULL
)
}
\arguments{
\item{cluster}{dataframe of variables in cluster space}

\item{settings}{list specifying parameters usually selected in the app}

\item{cov}{covariance matrix for space 1}

\item{covInv}{inverse covariance matrix for space 1}

\item{exp}{reference point in space 1}

\item{linked}{dataframe of variables in linked space}

\item{linked.cov}{covariance matrix for space 2}

\item{linked.covInv}{inverse covariance matrix for space 2}

\item{linked.exp}{reference point in space 2}

\item{user_dist}{user defined distances}

\item{getCoordsSpace1}{function to calculate coordinates in cluster space}

\item{getCoordsSpace2}{function to calculate coordinates in linked space}

\item{getScore}{function to calculate scores and bins}

\item{results}{an output of \code{makeResults()}, used to reduce computation when many plots are made.}
}
\value{
ggplot, plotly or detourr plot depending on settings$plotType
}
\description{
An interface to generate a specific graph seen when using the GUI.
Settings include: metric, linkage, k, plotType, for details
see the vignette on using this function.
}
\examples{
makePlots(
  cluster = Bikes$space1,
  settings = list(
    plotType = "WC", x = "hum", y = "temp", k = 4, metric = "euclidean",
    linkage = "ward.D2", WCa = 0.5, showalpha = TRUE
  ), cov = cov(Bikes$space1),
  linked = Bikes$space2, getScore = outsideScore(Bikes$other$res, "Residual")
)

makePlots(
  cluster = Bikes$space1,
  settings = list(
    plotType = "tour", k = 4, metric = "euclidean", linkage = "ward.D2",
    tourspace = "space1", colouring = "clustering", out_dim = 2, tour_path = "grand",
    display = "scatter", radial_start = NULL, radial_var = NULL, slice_width = NULL, seed = 2025
  ),
  cov = cov(Bikes$space1), linked = Bikes$space2,
  getScore = outsideScore(Bikes$other$res, "Residual")
)

}
