% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{makeResults}
\alias{makeResults}
\title{Generate Results for makePlots}
\usage{
makeResults(
  cluster,
  settings,
  cov = NULL,
  covInv = NULL,
  exp = NULL,
  linked = NULL,
  linked.cov = NULL,
  linked.covInv,
  linked.exp = NULL,
  user_dist = NULL,
  getCoordsSpace1 = normCoords,
  getCoordsSpace2 = normCoords,
  getScore = NULL
)
}
\arguments{
\item{cluster}{dataframe of variables in cluster space}

\item{settings}{list specifying parameters usually selected in the app}

\item{cov}{covariance matrix for space 1}

\item{covInv}{inverse covariance matrix for space 1}

\item{exp}{reference point in space 1}

\item{linked}{dataframe of variables in linked space}

\item{linked.cov}{covariance matrix for space 2}

\item{linked.covInv}{inverse covariance matrix for space 2}

\item{linked.exp}{reference point in space 2}

\item{user_dist}{user defined distances}

\item{getCoordsSpace1}{function to calculate coordinates in cluster space}

\item{getCoordsSpace2}{function to calculate coordinates in linked space}

\item{getScore}{function to calculate scores and bins}
}
\value{
list of results to be passed to makePlots
}
\description{
Settings are: metric, linkage, k. for details see the vignette on makePlots
}
\examples{
r <- makeResults(cluster = Bikes$space1, settings = list(k = 4,
   metric = "euclidean", linkage = "ward.D2"), cov = cov(Bikes$space1),
   linked = Bikes$space2, getScore = outsideScore(Bikes$other$res, "Residual"))
makePlots(cluster = Bikes$space1, settings = list(plotType = "Obs",
   x = "hum", y = "temp", obs = "A1"), cov = cov(Bikes$space1),
   linked = Bikes$space2, getScore = outsideScore(Bikes$other$res, "Residual"),
   results = r)

}
