% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{normCoords}
\alias{normCoords}
\title{Normalisation coordinates}
\usage{
normCoords(df, ...)
}
\arguments{
\item{df}{data frame}

\item{...}{other expected values of getCoords}
}
\value{
matrix with coordinate representation of all points
}
\description{
Normalisation coordinates
}
