% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{outsideScore}
\alias{outsideScore}
\title{Using externally computed score values}
\usage{
outsideScore(scores, scoreName = NULL)
}
\arguments{
\item{scores}{external scores to be passed to the app.}

\item{scoreName}{name for scores}
}
\value{
named list containing scores for use in pandemonium
}
\description{
Can be used as getScores input in pandemonium, to use score values
that are computed externally.
Returns scores values as the score, and bins computed
as below, between or above the first and third quartile.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
pandemonium(
  df = Bikes$space1, linked = Bikes$space2,
  getScore = outsidescore(Bikes$other$res, "Residual")
)
\dontshow{\}) # examplesIf}
}
