% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionReduction.R
\name{pca}
\alias{pca}
\title{Principal Component Analysis}
\usage{
pca(mat, ...)
}
\arguments{
\item{mat}{a coordinate matrix}

\item{...}{other parameters expected to be passed to dimReduction}
}
\value{
list containing a 2 x n matrix of reduced dimension data
}
\description{
Computes and returns first two principal components of a given matrix using
stats::prcomp
}
\examples{
head(pca(Bikes$space1)$Y)

}
