% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotChi2}
\alias{plotChi2}
\title{Plot chi2}
\usage{
plotChi2(wc, chi2, x, y, scoreName = NULL, cond = NULL)
}
\arguments{
\item{wc}{parameter values as matrix}

\item{chi2}{vector with chi2 values}

\item{x, y}{variables names (as string) to map to x and y axis}

\item{scoreName}{name for title}

\item{cond}{row numbers of points used for conditioning}
}
\value{
ggplot
}
\description{
Parameter values no longer need to be on a regular grid pattern for this plot.(modified)
}
\keyword{internal}
