% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCstat}
\alias{plotCstat}
\title{Plot selected cluster statistics}
\usage{
plotCstat(dist, fit, chivals, stat, kmax = 8)
}
\arguments{
\item{dist}{distances}

\item{fit}{result from hclust}

\item{chivals}{vector of chi2 values}

\item{stat}{cluster statistic to draw}

\item{kmax}{maximum number of clusters to appear in the plot}
}
\value{
ggplot
}
\description{
Plot selected cluster statistics
}
