% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotDimRed}
\alias{plotDimRed}
\title{Plot dimension reduction plot}
\usage{
plotDimRed(
  coord1,
  coord2,
  d_mat1,
  d_mat2,
  data,
  colouring,
  dimReduction,
  algorithm,
  group,
  score,
  user_group,
  pch
)
}
\arguments{
\item{coord1}{coordinates in space 1}

\item{coord2}{coordinates in space 2}

\item{d_mat1}{distance matrix in space 1}

\item{d_mat2}{distance matrix in space 2}

\item{data}{either "space1" or "space2"}

\item{colouring}{either "clustering", "user", "bins" or "score"}

\item{dimReduction}{function to calculate dimension reduction with $Y being the new n x 2 matrix}

\item{algorithm}{name for algorithm used for labeling plot}

\item{group}{grouping of points from clustering}

\item{score}{score values and bins}

\item{user_group}{user defined grouping}

\item{pch}{factor with 2 levels 1 will be plotted as a circle 2 will be plotted as an o}
}
\value{
plotly plot
}
\description{
Plot dimension reduction plot
}
