% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap with dendrogram}
\usage{
plotHeatmap(dat, fit, k, pal)
}
\arguments{
\item{dat}{coordinate representation of points}

\item{fit}{result from hclust}

\item{k}{number of clusters}

\item{pal}{color palette}
}
\value{
plot
}
\description{
Plot heatmap with dendrogram
}
