% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHist}
\alias{plotHist}
\title{Plot shadow Histograms}
\usage{
plotHist(dist, groups, n, bw.filt = NULL, gr.filt = NULL)
}
\arguments{
\item{dist}{distance matrix}

\item{groups}{group assignment vector}

\item{n}{number of points in distance matrix}

\item{bw.filt}{optional filter "within" or "between or vector of both for columns to display.}

\item{gr.filt}{optional filter of groups to show}
}
\value{
ggplot
}
\description{
Plot shadow Histograms
}
\keyword{internal}
