% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPC}
\alias{plotPC}
\title{Make parallel coordinate plot}
\usage{
plotPC(coord, groups, benchmarkIds, filt, c = T, s = T, a = 0.2, pal = NULL)
}
\arguments{
\item{coord}{coordinate representation of points}

\item{groups}{grouping from clustering is numeric or can be made numeric by as.numeric}

\item{benchmarkIds}{index values of benchmarks}

\item{filt}{filter of groups}

\item{c}{centre}

\item{s}{rescale (default=TRUE)}

\item{a}{alpha transarancy for drawing non-benchmark points (default=0.2)}

\item{pal}{pallete for colour assignment}
}
\value{
ggplot
}
\description{
Make parallel coordinate plot
}
