% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSigBin}
\alias{plotSigBin}
\title{Plot sigma bins in parameter space}
\usage{
plotSigBin(
  wc,
  interest,
  bmID,
  sigmabins,
  x,
  y,
  binName,
  cond = NULL,
  colourSet = "Set2"
)
}
\arguments{
\item{wc}{parameter values as matrix}

\item{interest}{logical vector showing that points are intersting}

\item{bmID}{index values for the benchmark points}

\item{sigmabins}{binning in sigma}

\item{x, y}{variables names (as string) to map to x and y axis}

\item{binName}{name for title}

\item{cond}{row numbers of points used for conditioning}

\item{colourSet}{RColorBrewer set for colouring}
}
\value{
ggplot
}
\description{
Parameter values no longer need to be on a regular grid pattern for this plot.(modified)
}
\keyword{internal}
