% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{pullCoordsNoCov}
\alias{pullCoordsNoCov}
\title{Generic Loss Function Coordinates}
\usage{
pullCoordsNoCov(df, cov, exp, ...)
}
\arguments{
\item{df}{data frame}

\item{cov}{covariance matrix}

\item{exp}{reference values}

\item{...}{other expected values of getCoords}
}
\value{
matrix with coordinate representation of all points
}
\description{
Generic Loss Function Coordinates
}
