% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{pullCoordsSqrt}
\alias{pullCoordsSqrt}
\title{Chi-Squared Loss Function Coordinates}
\usage{
pullCoordsSqrt(df, covInv, exp, ...)
}
\arguments{
\item{df}{data frame}

\item{covInv}{inverse covariance matrix}

\item{exp}{reference values}

\item{...}{other expected values of getCoords}
}
\value{
matrix with coordinate representation of all points
}
\description{
Computes coordinate values by comparing observed values
to the reference, using the square root inverse covariance matrix as when
computing the chi-squared loss.
}
\examples{
head(pullCoordsSqrt(
  Bikes$space2, solve(cov(Bikes$space2)),
  data.frame(value = colMeans(Bikes$space2))
))

}
