% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{rawCoords}
\alias{rawCoords}
\title{Raw coordinates}
\usage{
rawCoords(df, ...)
}
\arguments{
\item{df}{data frame}

\item{...}{other expected values of getCoords}
}
\value{
matrix with coordinate representation of all points
}
\description{
Returns the input data frame. This is used when other coordinate computations fail.
In general, scaling of the inputs is recommended before clustering.
}
\examples{
head(rawCoords(Bikes$space2))


}
