% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tour_coord}
\alias{tour_coord}
\title{Wrapper to prepare data for tour}
\usage{
tour_coord(coord, col, pch, addOrigin = FALSE)
}
\arguments{
\item{coord}{coordinate representation of all points}

\item{col}{color vector}

\item{pch}{pch vector}

\item{addOrigin}{set TRUE to add point through the origin}
}
\value{
list with formatted tour input
}
\description{
Wrapper to prepare data for tour
}
