% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tourGif}
\alias{tourGif}
\title{Make GIF of tour animation.}
\usage{
tourGif(coord, col, pch)
}
\arguments{
\item{coord}{coordinate representation of the points}

\item{col}{color vector according to group assignment}

\item{pch}{vector of plotting symbols}
}
\description{
Renders the tour animation, stored in file tour_animation.gif
}
