% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourMaker.R
\name{tourMaker}
\alias{tourMaker}
\title{function to make tours}
\usage{
tourMaker(
  coord1,
  coord2,
  group,
  score,
  user_group,
  tourspace,
  colouring,
  out_dim,
  tour_path,
  display,
  radial_start = NULL,
  radial_var = NULL,
  slice_width = NULL
)
}
\arguments{
\item{coord1}{coordinate matrix in space 1}

\item{coord2}{coordinate matrix in space 2}

\item{group}{grouping assignment}

\item{score}{score assignments}

\item{user_group}{user defined grouping}

\item{tourspace}{space to show tour of}

\item{colouring}{colouring to use in plot}

\item{out_dim}{dimension of output tour}

\item{tour_path}{tour path and type to use, one of ("grand","cmass","holes","lda","pda","dcor","spline","radial","anomaly")}

\item{display}{display type, one of ("scatter","slice")}

\item{radial_start}{projection to use as start of radial tour, one of ("random","cmass","holes","lda","pda","dcor","spline")}

\item{radial_var}{variable to remove by radial tour}

\item{slice_width}{width of slice}
}
\value{
detour
}
\description{
function to make tours
}
\keyword{internal}
