% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionReduction.R
\name{tSNE}
\alias{tSNE}
\title{t-Distributed Stochastic Neighbor Embedding}
\usage{
tSNE(dist, ...)
}
\arguments{
\item{dist}{a distance matrix}

\item{...}{other parameters expected to be passed to dimReduction}
}
\value{
list containing a n x 2 matrix of reduced dimension data in Y
}
\description{
Computes non-linear dimension reduction with Rtsne and default parameters.
}
\examples{
head(tSNE(getDists(Bikes$space1, "euclidean"))$Y)

}
