% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui}
\alias{ui}
\title{Generating layout for the graphical interface.}
\usage{
ui(params)
}
\arguments{
\item{params}{input parameters}
}
\value{
shiny ui
}
\description{
Generating layout for the graphical interface.
}
\keyword{internal}
