% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionReduction.R
\name{umap}
\alias{umap}
\title{Uniform Manifold Approximation and Projection Embedding}
\usage{
umap(dist, ...)
}
\arguments{
\item{dist}{a distance matrix}

\item{...}{other parameters expected to be passed to dimReduction}
}
\value{
list containing a 2 x n matrix of reduced dimension data
}
\description{
Computes non-linear dimension reduction with uwot and default parameters.
}
\examples{
head(umap(getDists(Bikes$space1, "euclidean"))$Y)

}
