% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{writeResults}
\alias{writeResults}
\title{Write coordinates and cluster assignment to a CSV file}
\usage{
writeResults(
  cluster,
  cov = NULL,
  covInv = NULL,
  exp = NULL,
  linked,
  linked.cov = NULL,
  linked.covInv = NULL,
  linked.exp = NULL,
  settings,
  filename,
  user_dist = NULL,
  getCoords.space1 = normCoords,
  getCoords.space2 = rawCoords
)
}
\arguments{
\item{cluster}{cluster space matrix}

\item{cov}{covariance matrix}

\item{covInv}{inverse covariance matrix}

\item{exp}{observable reference value (e.g. experimental measurement)}

\item{linked}{linked space matrix}

\item{linked.cov}{covariance matrix}

\item{linked.covInv}{inverse covariance matrix}

\item{linked.exp}{observable reference value (e.g. experimental measurement)}

\item{settings}{list specifying parameters usually selected in the app}

\item{filename}{path to write the results file to}

\item{user_dist}{input distance matrix (optional)}

\item{getCoords.space1}{function to calculate coordinates on clustering space}

\item{getCoords.space2}{function to calculate coordinates on linked space}
}
\value{
No return value, called for writing file
}
\description{
For working with the results outside the app.
Settings used: metric, linkage, k
}
\examples{
file <- tempfile()
writeResults(
  cluster = Bikes$space1, linked = Bikes$space2,
  settings = list(metric = "euclidean", linkage = "ward.D2", k = 4), filename = file
)
file.remove(file)

}
