% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bikes}
\alias{Bikes}
\title{bikes with activations}
\format{
a list of 4 dataframes
\describe{
\item{df}{dataframe 731 obs of 18 variables containing the entire bikes data set}
\item{space1}{dataframe 731 obs of 8 variables (cluster space)}
\item{space2}{dataframe 731 obs of 6 variables (linked space)}
\item{other}{dataframe 731 obs of 4 variables (other variables)}
}
}
\usage{
Bikes
}
\description{
data from Ursula
}
\keyword{datasets}
