% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{chi2Score}
\alias{chi2Score}
\title{Chi-squared scores function}
\usage{
chi2Score(space1, covinv, exp, ...)
}
\arguments{
\item{space1}{dataframe with variables in space1}

\item{covinv}{inverse covariance matrix from space1}

\item{exp}{reference point from space 1}

\item{...}{other expected values of getScore}
}
\value{
named list containing scores for use in pandemonium
}
\description{
Can be used as getScores input in pandemonium.
Returns chi-squared values as the score and sigma bins as the bins.
}
\examples{
chi2Score(
  Bikes$space1, solve(cov(Bikes$space1)),
  data.frame(value = colMeans(Bikes$space1))
)

}
