% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{chi2score}
\alias{chi2score}
\title{chi squared scores for the banomalies example}
\usage{
chi2score(space1, covinv, exp, space2, ...)
}
\arguments{
\item{space1}{dataframe with variables in space1}

\item{covinv}{inverse covariance matrix from space1}

\item{exp}{reference point from space 1}

\item{...}{other expected values of getScore}
}
\description{
used as getScores in newPandemonium
}
