% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{makePlots}
\alias{makePlots}
\title{Make plot as specified in settings}
\usage{
makePlots(
  pred,
  covInv,
  wc,
  exp,
  settings,
  user_coord = NULL,
  user_dist = NULL,
  c = NULL
)
}
\arguments{
\item{pred}{prediction matrix}

\item{covInv}{inverse covariance matrix}

\item{wc}{matrix specifying the model parameters (on a grid)}

\item{exp}{observable reference value (e.g. experimental measurement)}

\item{settings}{list specifying parameters usually selected in the app}

\item{user_coord}{input coordinate matrix (optional)}

\item{user_dist}{input distance matrix (optional)}

\item{c}{specification for conditioning}
}
\value{
ggplot
}
\description{
Used as interface to generate a specific graph seen when using the GUI.
Settings include: coord, useCov, metric, linkage, k, plotType
}
