% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{normCoords}
\alias{normCoords}
\title{Scaled coordinates}
\usage{
normCoords(df, ...)
}
\arguments{
\item{df}{data frame}

\item{...}{other expected values of getCoords}
}
\value{
matrix with coordinate representation of all points
}
\description{
Using scale to center and scale the coordinates.
}
\examples{
head(normCoords(Bikes$space2))

}
