% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{outsidescore}
\alias{outsidescore}
\title{outside score values}
\usage{
outsidescore(scores, scoreName = NULL)
}
\arguments{
\item{scores}{external scores to be passed to the app.}

\item{scoreName}{name for scores}
}
\description{
Used as getScores in pandemonium. Calculates scores not based on any data produced within pandemonium.
Returns scores values as the score and and lower,inter and upper quartiles as the bins.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pandemonium(df = Bikes$space1, space2 = Bikes$space2, getScore = outsidescore(Bikes$other$res,"Residual"))

\dontshow{\}) # examplesIf}
}
