% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{outsidescore}
\alias{outsidescore}
\title{outside score values}
\usage{
outsidescore(scores, scoreName = NULL)
}
\arguments{
\item{scores}{external scores to be passed to the app.}

\item{scoreName}{name for scores}
}
\description{
used as getScores in newPandemonium to calculate scores not based on any data within pandemonium
}
