% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotObs}
\alias{plotObs}
\title{Make coordinate plot}
\usage{
plotObs(coord, x, y, wc, obs, cond)
}
\arguments{
\item{coord}{coordinate representation of points}

\item{x, y}{variables names (as string) to map to x and y axis}

\item{wc}{parameter values as matrix}

\item{obs}{observable to plot}

\item{cond}{row numbers of points used for conditioning}
}
\value{
ggplot
}
\description{
Parameter values need to be on a regular grid for correct appearance of
this plot.
}
