% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotWC}
\alias{plotWC}
\title{Show clusters in parameter space}
\usage{
plotWC(
  wc,
  x,
  y,
  interest,
  bmID,
  col,
  cond = NULL,
  groups = NULL,
  pal = NULL,
  a = 0.2,
  showalpha = F
)
}
\arguments{
\item{wc}{parameter values as matrix}

\item{x, y}{variables names (as string) to map to x and y axis}

\item{interest}{index values for the intersting points}

\item{bmID}{index values of benchmarks}

\item{col}{color vector according to cluster assignment}

\item{cond}{row numbers of points used for conditioning}

\item{groups}{grouping assignments used to make alphahull}

\item{pal}{pallete used for group colouring of alphahull}

\item{a}{alpha value for alpha hull}

\item{showalpha}{boolean value to calculate and show alpha hulls}
}
\value{
ggplot
}
\description{
Parameter values no longer need to be on a regular grid pattern for this plot.(modified)
}
\keyword{internal}
