% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotWC}
\alias{plotWC}
\title{Show clusters in parameter space}
\usage{
plotWC(wc, x, y, sm, bf, benchmarkIds, col, cond = NULL)
}
\arguments{
\item{wc}{parameter values as matrix}

\item{x, y}{variables names (as string) to map to x and y axis}

\item{sm, bf}{index values for the SM and BF point}

\item{benchmarkIds}{index values of benchmarks}

\item{col}{color vector according to cluster assignment}

\item{cond}{row numbers of points used for conditioning}
}
\value{
ggplot
}
\description{
Parameter values need to be on a regular grid for correct appearance of
this plot.
}
