% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{pullCoords}
\alias{pullCoords}
\title{Chi-Squared Loss Function Coordinates}
\usage{
pullCoords(df, covInv, exp, ...)
}
\arguments{
\item{df}{data frame}

\item{covInv}{inverse covariance matrix}

\item{exp}{reference values}

\item{...}{other expected values of getCoords}
}
\value{
matrix with coordinate representation of all points
}
\description{
Chi-Squared Loss Function Coordinates
}
\examples{
pullCoords(Bikes$space2,solve(cov(Bikes$space2)),mean(Bikes$space2))


}
