% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{rawCoords}
\alias{rawCoords}
\title{Raw coordinates}
\usage{
rawCoords(df, ...)
}
\arguments{
\item{df}{data frame}

\item{...}{other expected values of getCoords}
}
\value{
matrix with coordinate representation of all points
}
\description{
returns df, used when other coordinates may fail.
It is not recommended that these coordinates be used as distances may not be logical.
}
\examples{
rawCoords(Bikes$space2)


}
