% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionReduction.R
\name{tSNE}
\alias{tSNE}
\title{t-Distributed Stochastic Neighbor Embedding Through Rtsne}
\usage{
tSNE(dist, ...)
}
\arguments{
\item{dist}{a distance matrix}

\item{...}{other parameters expected to be passed to dimReduction}
}
\value{
list containing a 2 x n matrix of reduced dimension data
}
\description{
t-Distributed Stochastic Neighbor Embedding Through Rtsne
}
\examples{
tSNE(getDists(Bikes$space1,"euclidean"))


}
