% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{writeResults}
\alias{writeResults}
\title{Write WC, coordinates and cluter assingment to a CSV file}
\usage{
writeResults(
  pred,
  covInv,
  wc,
  exp,
  settings,
  filename,
  user_coord = NULL,
  user_dist = NULL
)
}
\arguments{
\item{pred}{prediction matrix}

\item{covInv}{inverse covariance matrix}

\item{wc}{matrix specifying the model parameters (on a grid)}

\item{exp}{observable reference value (e.g. experimental measurement)}

\item{settings}{list specifying parameters usually selected in the app}

\item{filename}{path to write the results file to}

\item{user_coord}{input coordinate matrix (optional)}

\item{user_dist}{input distance matrix (optional)}
}
\description{
Settings include: coord, useCov, metric, linkage, k, plotType
}
