% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{writeResults}
\alias{writeResults}
\title{Write space2, coordinates and cluter assingment to a CSV file}
\usage{
writeResults(
  space1,
  cov,
  covInv,
  space2,
  exp,
  settings,
  filename,
  user_coord = NULL,
  user_dist = NULL,
  getCoords = normCoords
)
}
\arguments{
\item{space1}{cluster space matrix}

\item{cov}{covariance matrix}

\item{covInv}{inverse covariance matrix}

\item{space2}{space2 matrix}

\item{exp}{observable reference value (e.g. experimental measurement)}

\item{settings}{list specifying parameters usually selected in the app}

\item{filename}{path to write the results file to}

\item{user_coord}{input coordinate matrix (optional)}

\item{user_dist}{input distance matrix (optional)}

\item{getCoords}{function to calculate coordinates}
}
\description{
Settings include: metric, linkage, k, plotType
}
